WindowList_* functions are not services anymore; unneeded helpers removed
git-svn-id: http://svn.miranda-ng.org/main/trunk@14168 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/TrafficCounter/src/TrafficCounter.cpp
@@ -392,7 +392,7 @@ int TrafficCounter_Draw(HWND hwnd, HDC hDC)
 
 static void TC_AlphaText(HDC hDC, LPCTSTR lpString, RECT* lpRect, UINT format, BYTE ClistModernPresent)
 {
-	int nCount = mir_tstrlen( lpString );
+	int nCount = (int)mir_tstrlen( lpString );
 
 	if (ClistModernPresent)
 		AlphaText(hDC, lpString, nCount, lpRect, format, Traffic_FontColor);
